+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1] - (05/02/2020)
+
+### Added
+- .gitattributes file
+- \BristolSU\Support\User\UserRepository::getById method for retrieving a user by ID
+- \BristolSU\Support\User\UserRepository::setRememberToken to set a users remember token
+- \BristolSU\Support\User\User::getEmailForVerification to get the email to use for verification
+- \BristolSU\Support\User\User::getEmailForPasswordReset to get the email to use for password resets
+
 ## [2.0.1] - (05/02/2020)
 
 ### Added
@@ -54,7 +63,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/bristol-su/support/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/bristol-su/support/compare/v2.1...HEAD
+[2.1]: https://github.com/bristol-su/support/compare/v2.0.1...v2.1
 [2.0.1]: https://github.com/bristol-su/support/compare/v2.0...v2.0.1
 [2.0]: https://github.com/bristol-su/support/compare/v1.1...v2.0
 [1.1]: https://github.com/bristol-su/support/compare/v1.0.1...v1.1